// projeto/web/vite_react_p2/src/components/singles/DataTable/interfaces/DataTableClassNames.ts
export interface DataTableClassNames {
  container?: string;
  tableWrapper?: string;
  table?: string;
  thead?: string;
  theadRow?: string;
  th?: string;
  thSortable?: string;
  tbody?: string;
  tbodyContainer?: string;
  tbodyRow?: string;
  tbodyRowHover?: string;
  td?: string;
  emptyRow?: string;
  emptyCell?: string;
  footer?: string;
  footerText?: string;
  paginationContainer?: string;
  paginationSelect?: string;
  paginationButton?: string;
  paginationButtonDisabled?: string;
  paginationText?: string;
  loadingContainer?: string;
  loadingSpinner?: string;
  loadingText?: string;
}